Some modularity that will make updates easier.
I really don’t care for the way I have put together the file handling
stuff.  I will be back at a later date.
